+++ src/main/java/racingcar/utils/RandomGenerate.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 import java.util.Random;
