+++ src/test/java/racingcar/RacingRecordManagerTest.java
@@ -2,6 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.RacingRecord;
+import racingcar.utils.RacingRecordManager;
 
 import java.util.ArrayList;
 import java.util.List;
