+++ src/test/java/racingcar/RacingTest.java
@@ -3,6 +3,9 @@ package racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Racing;
+import racingcar.domain.RacingRecord;
 
 import java.util.List;
